Start working and refactoring
