+++ .rubocop.yml
@@ -1,0 +1,34 @@
+Rails:
+  Enabled: true
+
+AllCops:
+  TargetRubyVersion: 2.1
+
+Metrics/LineLength:
+  Max: 140
+
+# Offense count: 34
+# Configuration parameters: CountComments.
+Metrics/MethodLength:
+  Max: 60
+
+# Offense count: 11
+Metrics/CyclomaticComplexity:
+  Max: 13
+
+# Offense count: 10
+Metrics/PerceivedComplexity:
+  Max: 14
+
+# Offense count: 15
+Metrics/AbcSize:
+  Max: 25
+
+Style/Documentation:
+  Enabled: false
+
+Rails/HttpPositionalArguments:
+  Enabled: false
+
+Metrics/BlockLength:
+  Max: 50
